==> 1. gender_preprocessing.ipynb <==
import pandas as pd
import numpy as np
import re
from functools import reduce
import seaborn as sns
import missingno as msno
%matplotlib inline

import matplotlib.pyplot as plt
from scipy.stats import norm


# Set the figure size - handy for larger output
from matplotlib import pyplot as plt
plt.rcParams["figure.figsize"] = [6, 6]
# Set up with a higher resolution screen (useful on Mac)
%config InlineBackend.figure_format = 'retina'

# Import data

data_prep = pd.read_excel(f'D:\MARBURG VIRUS DISEASES/data/WOSCC/woscc_bon_23_02_verif_on_wos.xls')[
    ['UT (Unique WOS ID)','Funding Name Preferred', 'Open Access Designations','Publisher',#'Source Title',
    'Times Cited, All Databases','180 Day Usage Count', 'Since 2013 Usage Count','Publication Year','Affiliations',
    'Addresses','Reprint Addresses','Author Full Names']]
# create a columns of funding yes or no
data_prep['funding_yes'] = data_prep['Funding Name Preferred'].apply(lambda x: 1 if not pd.isnull(x) else 0)#.to_frame().drop('Funding Orgs', axis=1)     

data_prep['open_access_yes'] = data_prep['Open Access Designations'].apply(lambda x: 1 if not pd.isnull(x) else 0)#.to_frame().drop('Funding Orgs', axis=1)     

data_prep = data_prep.rename(columns={"UT (Unique WOS ID)": "wos_ID"})
data_prep = data_prep.applymap(lambda s:s.lower() if type(s) == str else s)

data_prep.head(3)

# data_prep['number_of_institution']

data_prep.shape

data_gender_dummies = pd.read_csv(f'../TABLES/author_gender_list_count_dummies.csv').drop(
                                ['Publication Year','Funding Orgs','funding_yes'], axis=1)
data_gender_dummies = data_gender_dummies.rename(columns={"UT (Unique WOS ID)": "wos_ID"})
data_gender_dummies["wos_ID"] = data_gender_dummies["wos_ID"].str.lower()
data_gender_dummies.head()

data_gender_dummies.shape

first_author_gender = pd.read_csv(f'../DATA_PREPROCESSING/first_author_gender.csv').drop(
                                ['Publication Year','Funding Orgs','funding_yes'], axis=1)
first_author_gender = first_author_gender.rename(columns={"UT (Unique WOS ID)": "wos_ID",'ga_gender':'first_author_gender'})
first_author_gender["wos_ID"] = first_author_gender["wos_ID"].str.lower()
# first_author_gender = pd.get_dummies(first_author_gender.set_index('wos_ID')['first_author_gender'], prefix='first_author').reset_index()
first_author_gender.head()

first_author_gender.shape

last_author_gender = pd.read_csv(f'../DATA_PREPROCESSING/last_author_gender.csv').drop(
                                ['Publication Year','Funding Orgs','funding_yes','Author Full Names'], axis=1)
last_author_gender = last_author_gender.rename(columns={"UT (Unique WOS ID)": "wos_ID",'gender':'last_author_gender'})
last_author_gender["wos_ID"] = last_author_gender["wos_ID"].str.lower()
# last_author_gender = pd.get_dummies(last_author_gender.set_index('wos_ID')['last_author_gender'], prefix='last_author').reset_index()
last_author_gender.head()

last_author_gender.shape

first_last_author_gender = pd.merge(first_author_gender, last_author_gender, on='wos_ID')
first_last_author_gender['first_last_author_gender'] = first_last_author_gender['first_author_gender']+'_'+first_last_author_gender['last_author_gender']
first_last_author_gender_dummy = pd.get_dummies(first_last_author_gender.set_index('wos_ID')['first_last_author_gender']).replace({0: np.nan})
first_last_author_gender_dummy.head(2)

first_last_author_gender_dummy.count()

last_author_gender_dummy = pd.get_dummies(last_author_gender.set_index('wos_ID')['last_author_gender'], prefix='last_author').reset_index()
last_author_gender_dummy.head()

first_author_gender_dummy = pd.get_dummies(first_author_gender.set_index('wos_ID')['first_author_gender'], prefix='first_author').reset_index()
first_author_gender_dummy.head()

data_topics = pd.read_pickle('D:\MARBURG VIRUS DISEASES\DATA_PREPROCESSING/data_prep_lemmatized_topics.pkl')[['Topic N°']].reset_index()
data_topics.head(2)

data_topics.shape

data_journal_old = pd.read_csv(f"../DATA_PREPROCESSING_old/journal_scope_and_impact_factors.csv")[['UT (Unique WOS ID)','Source Title','IF', 'JCI','percentageOAGold']];
data_journal_old = data_journal_old.rename(columns = {'UT (Unique WOS ID)':'wos_ID'})
data_journal_old['wos_ID'] = data_journal_old['wos_ID'].str.lower()
data_journal_old.head()

data_journal_old.shape

# open access

marburg_data_upw = pd.read_csv(f"../DATA_PREPROCESSING/marburg_data_upw.csv")[['UT (Unique WOS ID)','DOI','is_oa','oa_status']];
marburg_data_upw = marburg_data_upw.rename(columns = {'UT (Unique WOS ID)':'wos_ID'})
marburg_data_upw['wos_ID'] = marburg_data_upw['wos_ID'].str.lower()
marburg_data_upw.head()

marburg_data_upw.shape

marburg_data_upw.count()

# number of countries

country_data = pd.read_pickle('D:\MARBURG VIRUS DISEASES\DATA_PREPROCESSING/countries_split.pkl').drop('Addresses', axis=1)
country_data['number_of_country'] = country_data.notnull().sum(axis=1)
number_country = country_data[['number_of_country']]
number_country = number_country.reset_index().rename(columns = {'UT (Unique WOS ID)':'wos_ID'})
number_country['wos_ID'] = number_country['wos_ID'].str.lower()
number_country.head()

number_country.shape

# number of institutions
data_prep['number_of_institution'] =  pd.DataFrame(data_prep['Addresses'].str.split(';', expand=True)).count(axis='columns')
number_institution = data_prep[['wos_ID','number_of_institution']]
# number_institution = number_institution.reset_index().rename(columns = {'UT (Unique WOS ID)':'wos_ID'})
number_institution['wos_ID'] = number_institution['wos_ID'].str.lower()
number_institution.head()

number_institution.shape

data_countries_all = pd.read_pickle('D:\MARBURG VIRUS DISEASES\DATA_PREPROCESSING/data_concat_countries_and_wb_groups_income.pkl').drop(
                ['Addresses','funding_yes','open_access_yes','Funding Orgs', 'Open Access Designations',
                   'Times Cited, All Databases', '180 Day Usage Count','Since 2013 Usage Count', 'Publication Year',], 
                axis=1)
data_countries_all.head(2)

data_countries_all.shape

data_countries_all.columns[60:]

first_authors_countries = pd.read_csv('D:\MARBURG VIRUS DISEASES\DATA_PREPROCESSING/data_concat_countries_and_wb_groups_income_first_authors.csv').drop(['Unnamed: 0','Author Full Names','Addresses'], axis=1)
first_authors_countries = first_authors_countries.rename(columns = {'UT (Unique WOS ID)':'wos_ID'})
first_authors_countries['wos_ID'] = first_authors_countries['wos_ID'].str.lower()
first_authors_countries.head(2)

first_authors_countries.shape

last_authors_countries = pd.read_csv('D:\MARBURG VIRUS DISEASES\DATA_PREPROCESSING/data_concat_countries_and_wb_groups_income_last_authors.csv').drop(['Unnamed: 0','Author Full Names','Addresses'], axis=1)
last_authors_countries = last_authors_countries.rename(columns = {'UT (Unique WOS ID)':'wos_ID'})
last_authors_countries['wos_ID'] = last_authors_countries['wos_ID'].str.lower()
last_authors_countries.head(2)

last_authors_countries.shape

data_prep.columns

# no countries
dfs = [data_prep, data_gender_dummies, first_author_gender_dummy, last_author_gender_dummy, first_last_author_gender_dummy, data_topics, 
       data_journal_old, number_country,marburg_data_upw]

marburg_virus_data_clean_concat_and_indicator = reduce(lambda left,right: pd.merge(left,right,on=['wos_ID'], how='left'), dfs)

#save data
marburg_virus_data_clean_concat_and_indicator.to_csv(f'../DATA_PREPROCESSING/marburg_virus_data_clean_concat_and_indicator.csv')
marburg_virus_data_clean_concat_and_indicator.head(2)

# all countries
dfs = [data_prep, data_gender_dummies, first_author_gender_dummy, last_author_gender_dummy, first_last_author_gender_dummy, data_topics, 
       data_journal_old, number_country, marburg_data_upw,data_countries_all]

marburg_virus_data_clean_concat_and_indicator_all_countries = reduce(lambda left,right: pd.merge(left,right,on=['wos_ID'], how='left'), dfs)

#save data
marburg_virus_data_clean_concat_and_indicator_all_countries.to_csv(f'../DATA_PREPROCESSING/marburg_virus_data_clean_concat_and_indicator_all_countries.csv')
marburg_virus_data_clean_concat_and_indicator_all_countries.head(2)

# first author countries
dfs = [data_prep, data_gender_dummies, first_author_gender_dummy, last_author_gender_dummy, first_last_author_gender_dummy, data_topics, 
       data_journal_old, number_country, marburg_data_upw,first_authors_countries]

marburg_virus_data_clean_concat_and_indicator_first_author_countries = reduce(lambda left,right: pd.merge(left,right,on=['wos_ID'], how='left'), dfs)

#save data
marburg_virus_data_clean_concat_and_indicator_first_author_countries.to_csv(f'../DATA_PREPROCESSING/marburg_virus_data_clean_concat_and_indicator_first_author_countries.csv')
marburg_virus_data_clean_concat_and_indicator_first_author_countries.head(2)

# last author countries
dfs = [data_prep, data_gender_dummies, first_author_gender_dummy, last_author_gender_dummy, first_last_author_gender_dummy, data_topics, 
       data_journal_old, number_country, marburg_data_upw, last_authors_countries]

marburg_virus_data_clean_concat_and_indicator_last_author_countries = reduce(lambda left,right: pd.merge(left,right,on=['wos_ID'], how='left'), dfs)

#save data
marburg_virus_data_clean_concat_and_indicator_last_author_countries.to_csv(f'../DATA_PREPROCESSING/marburg_virus_data_clean_concat_and_indicator_last_author_countries.csv')
marburg_virus_data_clean_concat_and_indicator_last_author_countries.head(2)

marburg_virus_data_clean_concat_and_indicator.columns[:60]

marburg_virus_data_clean_concat_and_indicator.columns[:60:]

marburg_virus_data_clean_concat_and_indicator.columns[60:]

# limit data to after 2007
marburg_virus_data_clean_concat_and_indicator_2007 = marburg_virus_data_clean_concat_and_indicator[marburg_virus_data_clean_concat_and_indicator['Publication Year'] > 2006]
#save data
marburg_virus_data_clean_concat_and_indicator_2007.to_csv(f'../TABLES_gender/marburg_virus_data_clean_concat_and_indicator_2007.csv')
marburg_virus_data_clean_concat_and_indicator_2007.to_excel(f'../TABLES_gender/marburg_virus_data_clean_concat_and_indicator_2007.xls')
marburg_virus_data_clean_concat_and_indicator_2007.head(2)


marburg_virus_data_clean_concat_and_indicator_2007.shape

# limit data to after 2007
# all countries
marburg_virus_data_clean_concat_and_indicator_all_countries_2007 = marburg_virus_data_clean_concat_and_indicator_all_countries[marburg_virus_data_clean_concat_and_indicator_all_countries['Publication Year'] > 2006]
#save data
marburg_virus_data_clean_concat_and_indicator_all_countries_2007.to_csv(f'../TABLES_gender/marburg_virus_data_clean_concat_and_indicator_all_countries_2007.csv')
marburg_virus_data_clean_concat_and_indicator_all_countries_2007.to_excel(f'../TABLES_gender/marburg_virus_data_clean_concat_and_indicator_all_countries_2007.xls')
marburg_virus_data_clean_concat_and_indicator_all_countries_2007.head(2)


marburg_virus_data_clean_concat_and_indicator_all_countries_2007.shape

# limit data to after 2007
# first authors countries
marburg_virus_data_clean_concat_and_indicator_first_author_countries_2007 = marburg_virus_data_clean_concat_and_indicator_first_author_countries[marburg_virus_data_clean_concat_and_indicator_first_author_countries['Publication Year'] > 2006]
#save data
marburg_virus_data_clean_concat_and_indicator_first_author_countries_2007.to_csv(f'../TABLES_gender/marburg_virus_data_clean_concat_and_indicator_first_author_countries_2007.csv')
marburg_virus_data_clean_concat_and_indicator_first_author_countries_2007.to_excel(f'../TABLES_gender/marburg_virus_data_clean_concat_and_indicator_first_author_countries_2007.xls')
marburg_virus_data_clean_concat_and_indicator_first_author_countries_2007.head(2)


marburg_virus_data_clean_concat_and_indicator_first_author_countries_2007.shape

# limit data to after 2007
# last authors countries
marburg_virus_data_clean_concat_and_indicator_last_author_countries_2007 = marburg_virus_data_clean_concat_and_indicator_last_author_countries[marburg_virus_data_clean_concat_and_indicator_last_author_countries['Publication Year'] > 2006]
#save data
marburg_virus_data_clean_concat_and_indicator_last_author_countries_2007.to_csv(f'../TABLES_gender/marburg_virus_data_clean_concat_and_indicator_last_author_countries_2007.csv')
marburg_virus_data_clean_concat_and_indicator_last_author_countries_2007.to_excel(f'../TABLES_gender/marburg_virus_data_clean_concat_and_indicator_last_author_countries_2007.xls')
marburg_virus_data_clean_concat_and_indicator_last_author_countries_2007.head(2)


marburg_virus_data_clean_concat_and_indicator_last_author_countries_2007.shape

marburg_virus_data_clean_concat_and_indicator_last_author_countries_2007[['female_female', 'female_male',
       'female_unclassified', 'female_unknown', 'male_female', 'male_male',
       'male_unclassified', 'male_unknown', 'unclassified_female',
       'unclassified_male', 'unclassified_unclassified', 'unknown_female',
       'unknown_male', 'unknown_unknown']].count()

# marburg_virus_data_clean_concat_and_indicator_last_author_countries_2007['number_of_institution']
