==> 2. analysis_accessibility_open_access.ipynb <==
import pandas as pd
import numpy as np
import re
from functools import reduce
import seaborn as sns
import missingno as msno
%matplotlib inline

import matplotlib.pyplot as plt
from scipy.stats import norm


# Set the figure size - handy for larger output
from matplotlib import pyplot as plt
plt.rcParams["figure.figsize"] = [10, 6]
# Set up with a higher resolution screen (useful on Mac)
%config InlineBackend.figure_format = 'retina'

from scipy import stats #the only import you actually need for this
import numpy as np
import seaborn as sns
import pandas as pd
import matplotlib.pyplot as plt
%config InlineBackend.print_figure_kwargs={'facecolor' : "w"}

# function that tests distribution then picks test accordingly
def test_sig(x, y):
    if stats.shapiro(x).pvalue >= 0.05 and stats.shapiro(y).pvalue >= 0.05: #if they are normally distriuted
        print("t-test")
        return stats.ttest_ind(x,y)
    else:
        print("mann-whitney")
        return stats.mannwhitneyu(x, y)

## import data

data = pd.read_csv('D:\MARBURG VIRUS DISEASES\TABLES_gender/marburg_virus_data_clean_concat_and_indicator_2007.csv')#.drop(['Addresses'], axis=1)
data.head(2)

data.count().head(10)

# The accessibility of Marburg Virus Diseases publications (open access)

## proportion all

countries_oa_list_all = data.reset_index().groupby(['open_access_yes']).count().T
countries_oa_list_all.rename(columns = {0:'all_is_no_oa', 
                              1:'all_is_oa'}, inplace = True)

countries_oa_list_all['all_total'] = countries_oa_list_all['all_is_no_oa'] + countries_oa_list_all['all_is_oa']

for col in countries_oa_list_all.columns:
    countries_oa_list_all['%'+col] = countries_oa_list_all[col]/countries_oa_list_all['all_total']*100
    
countries_oa_list_all['%_diff(%all_is_oa-%all_is_no_oa)'] = countries_oa_list_all['%all_is_oa'] - countries_oa_list_all['%all_is_no_oa']

# save data
# countries_oa_list_all.to_csv(f'../TABLES/countries_all_oa_list.csv')

countries_oa_list_all

countries_oa_list_all[['all_is_no_oa','all_is_oa','all_total']].describe()

### significance

# first_ authors
first_male_OA = data.loc[(data['first_author_gender'] == 'male') & (data['number_of_author'] >1), 'open_access_yes'].to_list()
first_female_OA = data.loc[(data['first_author_gender'] == 'female') & (data['number_of_author'] >1), 'open_access_yes'].to_list()

dfs = [first_male_OA,first_female_OA]
for df in dfs:
    if len(df)>=3 :
        test_si = test_sig(first_male_OA, first_female_OA)
test_si

# last authors
last_male_OA = data.loc[(data['last_author_gender'] == 'male') & (data['number_of_author'] >1), 'open_access_yes'].to_list()
last_female_OA = data.loc[(data['last_author_gender'] == 'female') & (data['number_of_author'] >1), 'open_access_yes'].to_list()

dfs = [last_male_OA,last_female_OA]
for df in dfs:
    if len(df)>=3 :
        test_si = test_sig(last_male_OA, last_female_OA)
test_si

# solo authors
solo_male_OA = data.loc[(data['last_author_gender'] == 'male') & (data['number_of_author'] == 1), 'open_access_yes'].to_list()
solo_female_OA = data.loc[(data['last_author_gender'] == 'female') & (data['number_of_author'] == 1), 'open_access_yes'].to_list()

dfs = [solo_male_OA,solo_female_OA]
for df in dfs:
    if len(df)>=3:
        test_sig(solo_male_OA, solo_female_OA)

solo_female_OA

len(solo_female_OA)

solo_male_OA

len(solo_male_OA)

## proportion solo

data_solo = data.loc[data['number_of_author'] == 1]
data_solo.shape

data_solo['number_of_author'].value_counts()

countries_oa_list_gender = data_solo.reset_index().groupby(['open_access_yes']).count().T
countries_oa_list_gender.rename(columns = {0:'gender_is_no_oa', 
                              1:'gender_is_oa'}, inplace = True)

countries_oa_list_gender['gender_total'] = countries_oa_list_gender['gender_is_no_oa'] + countries_oa_list_gender['gender_is_oa']

for col in countries_oa_list_gender.columns:
    countries_oa_list_gender['%'+col] = countries_oa_list_gender[col]/countries_oa_list_gender['gender_total']*100
    
countries_oa_list_gender['%_diff(%gender_is_oa-%gender_is_no_oa)'] = countries_oa_list_gender['%gender_is_oa'] - countries_oa_list_gender['%gender_is_no_oa']

# save data
# countries_oa_list_fund.to_csv(f'../TABLES/countries_fund_oa_list.csv')

countries_oa_list_gender

## proportion no fund

data_no_fund = data.loc[data['funding_yes'] == 0]
data_no_fund.head(2)

countries_oa_list_no_fund = data_no_fund.reset_index().groupby(['open_access_yes']).count().T
countries_oa_list_no_fund.rename(columns = {0:'no_fund_is_no_oa', 
                              1:'no_fund_is_oa'}, inplace = True)

countries_oa_list_no_fund['no_fund_total'] = countries_oa_list_no_fund['no_fund_is_no_oa'] + countries_oa_list_no_fund['no_fund_is_oa']

for col in countries_oa_list_no_fund.columns:
    countries_oa_list_no_fund['%'+col] = countries_oa_list_no_fund[col]/countries_oa_list_no_fund['no_fund_total']*100
    
countries_oa_list_no_fund['%_diff(%no_fund_is_oa-%no_fund_is_no_oa)'] = countries_oa_list_no_fund['%no_fund_is_oa'] - countries_oa_list_no_fund['%no_fund_is_no_oa']

# save data
# countries_oa_list_no_fund.to_csv(f'../TABLES/countries_fund_no_oa_list.csv')

countries_oa_list_no_fund

## concat all dataframe and save

# concatenate dataframes on index
concat_countries_oa_list = pd.concat([countries_oa_list_all, 
                                        countries_oa_list_fund, 
                                        countries_oa_list_no_fund], 
                                       axis=1)
# save data
concat_countries_oa_list.to_csv(f'../TABLES/concat_countries_oa_list.csv')
concat_countries_oa_list
