==> 3. analysis_geographic_contribution.ipynb <==
import pandas as pd
import numpy as np
import re
from functools import reduce
import seaborn as sns
import missingno as msno
%matplotlib inline

import matplotlib.pyplot as plt
from scipy.stats import norm


# Set the figure size - handy for larger output
from matplotlib import pyplot as plt
plt.rcParams["figure.figsize"] = [10, 6]
# Set up with a higher resolution screen (useful on Mac)
%config InlineBackend.figure_format = 'retina'

## import data

data = pd.read_pickle('D:\MARBURG VIRUS DISEASES\DATA_PREPROCESSING/data_concat_countries_and_wb_groups_income.pkl')#.drop(['Addresses'], axis=1)

data.head(2)

data.count().head(10)

# The Geographic contribution to Marburg Virus Diseases publications

## Proportion of publications by countries

countries_list = data.count().to_frame().sort_values(by=0, ascending=False)
countries_list['%'] = countries_list[0]/countries_list.T['Addresses'][0]*100
# save data
countries_list.to_csv(f'../TABLES/countries_list.csv')

countries_list.head(60)

## Proportion of publications  by countries and funding


countries_fund_list = data.reset_index().groupby(['funding_yes']).count().T
countries_fund_list.rename(columns = {0:'no_fund', 
                              1:'fund'}, inplace = True)

countries_fund_list['total'] = countries_fund_list['no_fund'] + countries_fund_list['fund']

for col in countries_fund_list.columns:
    countries_fund_list['%_'+col] = countries_fund_list[col]/countries_fund_list['total']*100
    
countries_fund_list['%_diff(%_fund-%_no_fund)'] = countries_fund_list['%_fund'] - countries_fund_list['%_no_fund']

# save data
countries_fund_list.to_csv(f'../TABLES/countries_fund_list.csv')

countries_fund_list

countries_list = [ 'Afghanistan', 'United States', 'Angola', 'Argentina', 'Australia', 'Austria', 'Bangladesh', 'Belgium', 'Belize', 'Brazil', 'Bulgaria', 'Burkina Faso',
 'Canada', 'Central African Republic', 'Colombia', 'Costa Rica', "Cote d'Ivoire", 'Croatia', 'Czech Republic',
 'DR Congo', 'Denmark', 'Ecuador', 'Egypt', 'United Kingdom', 'Germany', 'Finland', 'France', 'Gabon', 'Ghana', 'Greece', 'Guinea', 'Honduras', 'Hungary',
 'India', 'Indonesia', 'Iran', 'Iraq', 'Israel', 'Italy', 'Japan', 'Jordan', 'Kazakhstan', 'Kenya', 'Lebanon', 'Malawi', 'Malaysia', 'Mali', 'Morocco',
 'Nepal', 'Netherlands', 'New Zealand', 'Nigeria', 'Norway', 'Pakistan', 'Palestine', 'Panama', 'China', 'Peru', 'Philippines', 'Poland', 'Portugal',
 'Congo Republic', 'Russia', 'Saudi Arabia', 'Senegal', 'Serbia', 'Sierra Leone', 'Singapore', 'Slovakia', 'South Africa', 'South Korea', 'Spain', 'Sweden',
 'Switzerland', 'Tanzania', 'Thailand', 'Turkey', 'United Arab Emirates', 'Uganda', 'Ukraine', 'Venezuela', 'Vietnam', 'Zambia', 'Zimbabwe', 'Sudan']

region_wb = ['wb_east_asia_and_pacific', 'wb_europe_and_central_asia', 'wb_latin_america_and_the_caribbean', 'wb_middle_east_and_north_africa', 
               'wb_middle_east_and_north_africa_africa', 'wb_middle_east_and_north_africa_others', 'wb_north_america', 'wb_south_asia', 'wb_sub_saharan_africa',
                 ]

income_wb = ['wb_low_income_economies',  'wb_lower_middle_income_economies', 
               'wb_upper_middle_income_economies', 'wb_high_income_economies'
                  ]

countries_fund_list.reset_index()[countries_fund_list.reset_index()['index'].isin(countries_list)].describe()

countries_fund_list.reset_index()[countries_fund_list.reset_index()['index'].isin(region_wb)].describe()

countries_fund_list.reset_index()[countries_fund_list.reset_index()['index'].isin(income_wb)].describe()
