==> 3. analysis_scientific_collab.ipynb <==
import pandas as pd
import numpy as np
import re
from functools import reduce
import seaborn as sns
import missingno as msno
%matplotlib inline

import matplotlib.pyplot as plt
from scipy.stats import norm


# Set the figure size - handy for larger output
from matplotlib import pyplot as plt
plt.rcParams["figure.figsize"] = [10, 6]
# Set up with a higher resolution screen (useful on Mac)
%config InlineBackend.figure_format = 'retina'

# import data

data = pd.read_pickle('D:\MARBURG VIRUS DISEASES\DATA_PREPROCESSING/data_concat_countries_and_wb_groups_income.pkl')

data.head(1)

country_data = pd.read_pickle('D:\MARBURG VIRUS DISEASES\DATA_PREPROCESSING/countries_split.pkl').drop('Addresses', axis=1)
country_data.head(1)

data_prep = pd.read_excel(f'D:\MARBURG VIRUS DISEASES/data/WOSCC/woscc_bon_23_02_verif_on_wos.xls')
data_prep = data_prep.rename(columns={"UT (Unique WOS ID)": "wos_ID"})
data_prep = data_prep.applymap(lambda s:s.lower() if type(s) == str else s)
#data_prep = data_prep.drop_duplicates('wos_ID')
data_prep.head(1)

data_prep.columns

# Extract indicators

## Extract number of countries

country_data['number_of_country'] = country_data.notnull().sum(axis=1)
country_data.head()

## Extract number of institution

# number of institutions
data_prep['number_of_institution'] =  pd.DataFrame(data_prep['Addresses'].str.split(';', expand=True)).count(axis='columns')
data_prep['number_of_institution']#.rename(columns={0: "Number_of_author_per_study"})

## Extract number of authors

# number of authors
data_prep['number_of_author'] =  pd.DataFrame(data_prep['Author Full Names'].str.split(';', expand=True)).count(axis='columns')#.rename(columns={0: "Number_of_author_per_study"})
data_prep['number_of_author']

## Extract types of collaboration

country_data['number_of_country'].value_counts()

# Create new column based 
country_data['collab_types'] = country_data['number_of_country'].apply(lambda x: 'Nat. collab.' if x <= 1 else 'Int. collab.')
country_data['collab_types']

country_data['collab_types'].value_counts()

data_region= data[['north_africa', 'eastern_africa', 'Southern_africa', 'Western_africa', 'Central_africa', 'non_african_countries', 
            'wb_east_asia_and_pacific', 'wb_europe_and_central_asia', 'wb_latin_america_and_the_caribbean', 
            'wb_middle_east_and_north_africa', 'wb_north_america', 'wb_south_asia', 'wb_sub_saharan_africa'
             ]]

data_region.notnull().sum(axis=1).value_counts()
